main: ensure port and PORT vars are always defined in the apprc file
We tried to fix a compatibility issue with tsurud 0.12.0 by not writing
environment variables to apprc, but as reported in tsuru/tsuru#995,
tsurud doesn't manage the port/PORT environment variable.

In order to properly handle the port environment variable, we need to
include it in the variables exported in the container on the
CreateContaner call. tsurud 0.12.2 will handle that, but we need to
ensure that we can still deploy code to tsurud 0.12.0 and 0.12.1, just
to give some for upgrading the version of the server.

So, in order to fix that, we always generate the apprc file, and if it's
tsurud 0.12.0 or greater, the file will contain only the port/PORT
environment variables. This is far from ideal, but handles the issue
without forcing users to update the API (but, of course, they're still
required to update the platforms).

Again, users will not need this in the future, as tsurud 0.12.2 should
fix the environment variables bugs introduced in tsurud 0.12.0, but
tsuru-unit-agent must work across multiple versions of tsurud.

It's what they say: what you see as an workaround, I see as an elegant
design solution.